проверка на то что партия не удалена

+++ party/views/management/staff/join_in_party.py
@@ -20,26 +20,30 @@ def join_in_party(request, plr_pk, pty_pk):
         logged_player = Player.get_instance(pk=plr_pk)
         # если игрок - беспартийный
         if not logged_player.party:
-            # если партия является открытой
-            if Party.objects.get(pk=pty_pk).type == 'op':
-                # удаляем все его заявки в другие партии
-                PartyApply.objects.filter(player=logged_player).delete()
-                # задаем его новую партию
-                logged_player.party = Party.objects.get(pk=pty_pk)
-                # задаем ему должность нового игрока в партии
-                # "базовая должность без лидерки - это 'новичок' "
-                party_newbie_post = PartyPosition.objects.get(based=True, party=Party.objects.get(pk=pty_pk),
-                                                              party_lead=False)
-                logged_player.party_post = party_newbie_post
-                # Сохраняем это состояние
-                logged_player.save()
-                # Логировние: создаем запись о вступлении
-                party_log = MembershipLog(player=logged_player, dtime=timezone.now(),
-                                          party=Party.objects.get(pk=pty_pk))
-                party_log.save()
+            # если партия существует и не удалена
+            if Party.objects.filter(pk=pty_pk, deleted=False).exists():
+                # если партия является открытой
+                if Party.objects.get(pk=pty_pk).type == 'op':
+                    # удаляем все его заявки в другие партии
+                    PartyApply.objects.filter(player=logged_player).delete()
+                    # задаем его новую партию
+                    logged_player.party = Party.objects.get(pk=pty_pk)
+                    # задаем ему должность нового игрока в партии
+                    # "базовая должность без лидерки - это 'новичок' "
+                    party_newbie_post = PartyPosition.objects.get(based=True, party=Party.objects.get(pk=pty_pk),
+                                                                  party_lead=False)
+                    logged_player.party_post = party_newbie_post
+                    # Сохраняем это состояние
+                    logged_player.save()
+                    # Логировние: создаем запись о вступлении
+                    party_log = MembershipLog(player=logged_player, dtime=timezone.now(),
+                                              party=Party.objects.get(pk=pty_pk))
+                    party_log.save()
 
-                # Обновляем страницу
-                return redirect('party')
+                    # Обновляем страницу
+                    return redirect('party')
+                else:
+                    return redirect('party')
             else:
                 return redirect('party')
         else:
